fix: ポケモンの種族値取得処理をStatusコンポーネントの外に出す
