+++ src/pages/index.page.tsx
@@ -1,11 +1,10 @@
-import { Box, Center, HStack, Skeleton, Spacer } from "@chakra-ui/react";
+import { Box, Center, HStack, Spacer } from "@chakra-ui/react";
 import { Suspense } from "react";
 
-import { CURRENT, ADJUSTED } from "@/constants";
 import { Differences } from "@/features/differences/components";
 import { FormImages } from "@/features/forms/components";
 import { SearchForm } from "@/features/search/components";
-import { Status } from "@/features/status/components";
+import { StatusTableList } from "@/features/status/components";
 
 export function Home() {
   return (
@@ -17,20 +16,7 @@ export function Home() {
         </Suspense>
       </HStack>
       <Center width="100vw">
-        <HStack spacing="32px">
-          <Suspense fallback={<Skeleton />}>
-            <Status
-              label="現在のステータス"
-              statusType={CURRENT}
-            />
-          </Suspense>
-          <Suspense fallback={<Skeleton />}>
-            <Status
-              label="調整後のステータス"
-              statusType={ADJUSTED}
-            />
-          </Suspense>
-        </HStack>
+        <StatusTableList />
       </Center>
       <Box width="100%">
         <Differences />
